Add Gaussian components (closes #1)
I spent a lot of time trying to create framework that was component 'agnostic'. My early attempts were to use multiple dispatch on the component type, but this was much slower for GPU programming.

In the end, I'm adhering to the Rule of Three - with just Gaussian and point sources it's too early to refactor out common code. And besides, a point source is just a special type of Gaussian - so that's exactly what I've opted for.

There's a lot of knowledge spread through the code about this - far from ideal - and so if we ever integrate eg. shapelets, first priority will be to refactor this knowledge into the components themselves.